Added class types, beginnings of train functionality, and other stuff
Added class types because I learned the hard way that types are important!

Added train_pokemon functionality, which allows for finding optimal evs

I want to add features to allow for finding optimal evs for neutral
and supereffective hits, as well as for multiple pokemon kos.

It would be cool to also solve for 2hko, 3hko as well.

: